split out asset scaling
